+++ App/src/components/Request.js
@@ -2,40 +2,55 @@
 import React from "react";
 import { jsx } from "@emotion/core";
 
-function Request({request, findName, handleRequestSchedule, Turn}) {
+function Request({request, frontdesks, handleRequestSchedule}) {
   return (
     <li
-    key={request.id}
-    css={{
-      padding: "20px",
-      border: "1px solid black",
-      display: "flex",
-      flexDirection: "column",
-      "&:hover": {
-        cursor: "pointer"
-      }
-    }}
-  >
-    <p>#{request.id}</p>
-    <div>
+            key={request.id}
+            css={{
+              padding: "20px",
+              border: "1px solid black",
+              display: "flex",
+              flexDirection: "column",
+              "&:hover": {
+                cursor: "pointer"
+              }
+            }}
+          >
+      <p>#{request.id}</p>
       <span>
-        FrontDesk <b>{findName(request.requester_id)}</b> of
-        workshift <b>{Turn[request.current_Shift_id]} </b>
-        want to change workshift with FrontDesk
+        FrontDesk{" "}
+        {
+          frontdesks.find(frontdesk => frontdesk.id === request.requester_id)
+            .name
+        }{" "}
+        of workshift{" "}
+        {request.current_Shift_id === 4
+          ? "OFF"
+          : request.current_Shift_id === 1
+          ? "Morning"
+          : request.current_Shift_id === 2
+          ? "Afternoon"
+          : "Night"}{" "}
+        want to change workshift with FrontDesk{" "}
+        {
+          frontdesks.find(frontdesk => frontdesk.id === request.requested_id)
+            .name
+        }{" "}
+        of workshift{" "}
+        {request.requested_Shift_id === 4
+          ? "OFF"
+          : request.requested_Shift_id === 1
+          ? "Morning"
+          : request.requested_Shift_id === 2
+          ? "Afternoon"
+          : "Night"}
       </span>
-      <span>
-        <b> {findName(request.requested_id)} </b>
-        of workshift <b>{Turn[request.requested_Shift_id]} </b>
-      </span>
-    </div>
-    <div css={{ alignSelf: "flex-end" }}>{request.status}</div>
-    <button onClick={() => handleRequestSchedule(request.id)}>
-      See Schedule Request{" "}
-    </button>
-  </li>
+      <div css={{ alignSelf: "flex-end" }}>{request.status}</div>
+      <button onClick={() => handleRequestSchedule(request.id)}>
+        See Schedule Request{" "}
+      </button>
+    </li>
   );
 }
 
 export default Request;
-
-
